+++ src/components/container/MovieContainer.js
@@ -5,16 +5,16 @@ import Movie from '../presentation/Movie';
 
 const MovieContainer = (props) => {
   const { movies, filter } = props;
-  // console.log(movies);
 
-  const filterMovies = (movies, filter) => (filter === 'All' ? movies.Search : movies.Search.filter((movie) => movie.Type === filter));
+  const filterFilms = (movies, filter) => (filter === 'All' ? movies.Search
+    : movies.Search.filter((mov) => mov.Type === filter));
 
-  const filteredMovies = filterMovies(movies, filter);
-  // console.log(movies);
+  const AlreadyFiltered = filterFilms(movies, filter);
 
+  console.log(movies);
   let movieBody = '';
   movieBody = movies.Response === 'True'
-    ? filteredMovies.map((movie) => <Movie key={movie.Poster} movie={movie} />) : '';
+    ? AlreadyFiltered.map((movie) => <Movie key={movie.Poster} movie={movie} />) : '';
   return (
     <div className="row">
       {movieBody}
