+++ src/components/presentation/App.js
@@ -1,6 +1,7 @@
 import Navbar from './NavBar';
 import Footer from './Footer';
 import Routes from '../../routes/routes';
+import '../styles/App.css';
 
 const App = () => (
   <div className="app-div">
